refactor: split into chunk files
close #201
