+++ package.json
@@ -3,9 +3,9 @@
   "version": "0.10.3",
   "type": "module",
   "description": "Perform async work synchronously in Node.js using `worker_threads` with first-class TypeScript support.",
-  "repository": "git+https://github.com/un-ts/synckit.git",
-  "author": "JounQin (https://www.1stG.me) <admin@1stg.me>",
-  "funding": "https://opencollective.com/unts",
+  "repository": "https://github.com/un-ts/synckit.git",
+  "author": "JounQin <admin@1stg.me> (https://www.1stG.me)",
+  "funding": "https://opencollective.com/synckit",
   "license": "MIT",
   "packageManager": "yarn@4.8.1",
   "engines": {
@@ -15,6 +15,10 @@
   "module": "./lib/index.js",
   "exports": {
     "import": {
+      "types": "./lib/index.d.ts",
+      "default": "./lib/index.js"
+    },
+    "module-sync": {
       "types": "./lib/index.d.ts",
       "default": "./lib/index.js"
     },
@@ -50,6 +54,7 @@
     "build": "concurrently -r 'yarn:build:*'",
     "build:r": "r -f cjs",
     "build:ts": "tsc -p src",
+    "format": "prettier --write .",
     "lint": "concurrently -r 'yarn:lint:*'",
     "lint:es": "eslint . --cache --max-warnings 10",
     "lint:tsc": "tsc --noEmit",
@@ -64,14 +69,13 @@
     "tslib": "^2.8.1"
   },
   "devDependencies": {
-    "@1stg/common-config": "^11.0.5",
+    "@1stg/common-config": "^11.1.1",
     "@changesets/changelog-github": "^0.5.1",
     "@changesets/cli": "^2.28.1",
     "@commitlint/cli": "^19.8.0",
     "@pkgr/rollup": "^6.0.1",
-    "@swc-node/register": "^1.10.10",
     "@types/jest": "^29.5.14",
-    "@types/node": "^22.13.11",
+    "@types/node": "^22.13.16",
     "@unts/patch-package": "^8.1.1",
     "clean-pkg-json": "^1.2.1",
     "concurrently": "^9.1.2",
@@ -81,15 +85,17 @@
     "eslint": "^9.23.0",
     "jest": "^29.7.0",
     "lint-staged": "^15.5.0",
-    "make-synchronized": "^0.7.0",
-    "node-gyp": "^11.1.0",
+    "make-synchronized": "^0.7.2",
+    "node-gyp": "^11.2.0",
     "prettier": "^3.5.3",
     "simple-git-hooks": "^2.12.1",
+    "size-limit": "^11.2.0",
+    "size-limit-preset-node-lib": "^0.3.0",
     "sync-threads": "^1.0.1",
     "synckit": "link:.",
-    "tinyexec": "^1.0.0",
+    "tinyexec": "^1.0.1",
     "ts-expect": "^1.3.0",
-    "ts-jest": "^29.2.6",
+    "ts-jest": "^29.3.1",
     "ts-node": "^10.9.2",
     "tsx": "^4.19.3",
     "type-coverage": "^2.29.7",
@@ -116,16 +122,13 @@
     ],
     "moduleNameMapper": {
       "^(\\.{1,2}/.*)\\.js$": "$1",
-      "^synckit$": "<rootDir>/src"
+      "^synckit$": "<rootDir>/src/index.ts"
     },
     "transform": {
       "^.+\\.tsx?$": [
         "ts-jest",
         {
-          "useESM": true,
-          "tsconfig": {
-            "importHelpers": false
-          }
+          "useESM": true
         }
       ]
     }
